+++ tests/conftest.py
@@ -1,25 +1,25 @@
 import pytest
-from app.app import app as flask_app  # Ensure correct import
-from app.models import db  # Import database models
+from app.app import app as flask_app  # Changed import to use app/app.py directly
+from app.models.models import db
 
 @pytest.fixture
 def app():
     """Flask app configured for testing with an in-memory database."""
     flask_app.config.update({
         "TESTING": True,
-        "SQLALCHEMY_DATABASE_URI": "sqlite:///:memory:",  # ✅ Use in-memory SQLite instead of real DB
+        "SQLALCHEMY_DATABASE_URI": "sqlite:///:memory:",
         "SQLALCHEMY_TRACK_MODIFICATIONS": False,
-        "WTF_CSRF_ENABLED": False  # Disable CSRF for easier testing
+        "WTF_CSRF_ENABLED": False
     })
 
     with flask_app.app_context():
-        db.create_all()  # ✅ Create fresh tables in SQLite memory
+        db.create_all()
 
-    yield flask_app  # Provide app to tests
+    yield flask_app
 
     with flask_app.app_context():
         db.session.remove()
-        db.drop_all()  # ✅ Clean up after tests
+        db.drop_all()
 
 @pytest.fixture
 def client(app):
